Refactor book methods into BookList class
